+++ python/cudf/cudf/core/column/string.py
@@ -1084,28 +1084,6 @@ class StringColumn(column.ColumnBase):
         return len(self.unique()) == len(self)
 
     @property
-    def is_monotonic_increasing(self):
-        if not hasattr(self, "_is_monotonic_increasing"):
-            if self.nullable and self.has_nulls:
-                self._is_monotonic_increasing = False
-            else:
-                self._is_monotonic_increasing = libcudf.issorted.issorted(
-                    columns=[self]
-                )
-        return self._is_monotonic_increasing
-
-    @property
-    def is_monotonic_decreasing(self):
-        if not hasattr(self, "_is_monotonic_decreasing"):
-            if self.nullable and self.has_nulls:
-                self._is_monotonic_decreasing = False
-            else:
-                self._is_monotonic_decreasing = libcudf.issorted.issorted(
-                    columns=[self], descending=[1]
-                )
-        return self._is_monotonic_decreasing
-
-    @property
     def __cuda_array_interface__(self):
         raise NotImplementedError(
             "Strings are not yet supported via `__cuda_array_interface__`"