+++ CHANGELOG.md
@@ -129,6 +129,7 @@
 - PR #4244 Port nvstrings Substring Gather/Scatter functions to cuDF Python/Cython
 - PR #4280 Port nvstrings Numeric Handling functions to cuDF Python/Cython
 - PR #4336 Move a bunch of internal nvstrings code to use native StringColumns
+- PR #4166 Port `is_sorted.pyx` to use libcudf++ APIs
 
 ## Bug Fixes